requirements.txt
streamlit
python-dotenv
langchain
langchain-community
huggingface_hub
PyPDF2
